// src/query/Enum.ts
export enum ExprTypeEnum {
    and = "AND",
    or = "OR",
    equal = "EQUAL",
    gt = "GT",
    gte = "GTE",
    lt = "LT",
    lte = "LTE",
    in = "IN",
    nin = "NIN",
    regex = "REGEX",
}
/** 排序枚举 */
export enum QuerySortEnum {
    /** 倒序 */
    desc = "DESC",
    /** 正序 */
    asc = "ASC",
}